ControllerAdding (#24)
* UserInput event-handlers refactoring (stage one).

* Refactoring of Controllers input functionality

* Fix "unnesessary empty line" and white_space in line #200 in UserInput.cpp

* Add "Menu" and "Pause" buttons functionality for controllers.

* Fix for issue with moving player while menu is active and refactoring of userinput subscriptions.

* Fix for issue with moving enemies and paused players while Menu is opened.

* Revert "Fix for issue with moving player while menu is active and refactoring of userinput subscriptions."

This reverts commit 90250c2c38b98dc7415ca58ef246491be1051166.

* restore after revert

* WIP: state manager and pause dispaying

* fix and some revert after sub merge

* fix pause switch on menu open, reformating and cleanup

* state manager adding gameover and pause refactoring, file renaming

* Made new way to handle "GameOver" state into the SateManager

* add preUpdate and postUpdate stage

* add PreDraw and PostDraw and PreDrawUserInterface and PostDrawUserInterface stages and its fix bugs with overlaping, cleanup game success

* add FramePerSecondManager and refactor GameSuccess

* fix cmake build and unit tests

* refatoring animation manager, cleanup dependencies

* add spawnManager and fix unit tests, refactoring obstacle spawner and map loading

* resolve review issues

* add renderManager, refactor rendering logic so sld rendered in one class, cleanup unused interfaces

* fix color in menu white->cyan and refactoring int color to unsigned int, reformat

* Add WinCondition and visual placeholder for it.

* WinCondition refactoring to be alligned with RenderManager.

* fix recenter health bar when pickup star bonus and fix some warnings

* fix for fps dispaying and fps drop, cleanup

* refactoring fps count, fix cmake build, deltatime now double, cleanup

* Refactor of wincondition functionality.

* renamings, reformat

* add bonus(start, caliber, tank) network handling, cleanup outdated server stuff and unify parameter usage

* refactor statistics, fix unsubscribe, cleanup

* fix unsubscribe and reorder, cleanup

* reformat

* fix build, after removing SDL_Log dependency

* Remake "GameOver" logic in "GameWon"-like style.

* Deletion of unnecessary GameOver-related conditions in StateManager.

* Refactoring of Win/Loose functionality.

* One more refactoring for Win/Loose functionality.

* refactor win\lose condition

---------

+++ headers/behavior/MoveLikeTankBeh.h
@@ -27,14 +27,14 @@ class MoveLikeTankBeh final : public IMoveBeh
 
 	void HandleBonusPickUp(const std::shared_ptr<BaseObj>& object) const;
 
-	[[nodiscard]] bool IsCanMove(float deltaTime) const override;
+	[[nodiscard]] bool IsCanMove(double deltaTime) const override;
 	[[nodiscard]] float FindMinDistance(const std::vector<std::shared_ptr<BaseObj>>& objects,
 	                                    const std::function<float(const std::shared_ptr<BaseObj>&)>& sideDiff) const;
 
-	[[nodiscard]] bool MoveLeft(float deltaTime) override;
-	[[nodiscard]] bool MoveRight(float deltaTime) override;
-	[[nodiscard]] bool MoveUp(float deltaTime) override;
-	[[nodiscard]] bool MoveDown(float deltaTime) override;
+	[[nodiscard]] bool MoveLeft(double deltaTime) override;
+	[[nodiscard]] bool MoveRight(double deltaTime) override;
+	[[nodiscard]] bool MoveUp(double deltaTime) override;
+	[[nodiscard]] bool MoveDown(double deltaTime) override;
 
 public:
 	MoveLikeTankBeh(ObjRectangle& rect, Direction& dir, float& speed, buuid& uuid, UPoint& windowSize,
@@ -43,7 +43,7 @@ public:
 
 	~MoveLikeTankBeh() override = default;
 
-	[[nodiscard]] bool Move(float deltaTime) override;
+	[[nodiscard]] bool Move(double deltaTime) override;
 
-	[[nodiscard]] std::vector<Direction> GetFreePathSides(float deltaTime) const override;
+	[[nodiscard]] std::vector<Direction> GetFreePathSides(double deltaTime) const override;
 };
