+++ headers/interfaces/IMoveBeh.h
@@ -8,18 +8,18 @@ class BaseObj;
 
 class IMoveBeh
 {
-	[[nodiscard]] virtual bool IsCanMove(float deltaTime) const = 0;
+	[[nodiscard]] virtual bool IsCanMove(double deltaTime) const = 0;
 
 protected:
-	[[nodiscard]] virtual bool MoveLeft(float deltaTime) = 0;
-	[[nodiscard]] virtual bool MoveRight(float deltaTime) = 0;
-	[[nodiscard]] virtual bool MoveUp(float deltaTime) = 0;
-	[[nodiscard]] virtual bool MoveDown(float deltaTime) = 0;
+	[[nodiscard]] virtual bool MoveLeft(double deltaTime) = 0;
+	[[nodiscard]] virtual bool MoveRight(double deltaTime) = 0;
+	[[nodiscard]] virtual bool MoveUp(double deltaTime) = 0;
+	[[nodiscard]] virtual bool MoveDown(double deltaTime) = 0;
 
 public:
 	virtual ~IMoveBeh() = default;
 
-	[[nodiscard]] virtual bool Move(float deltaTime) = 0;
+	[[nodiscard]] virtual bool Move(double deltaTime) = 0;
 
-	[[nodiscard]] virtual std::vector<Direction> GetFreePathSides(float deltaTime) const = 0;
+	[[nodiscard]] virtual std::vector<Direction> GetFreePathSides(double deltaTime) const = 0;
 };
